Fix: Alert 로직을 Utility로 분리
